# src/config/swagger.json
{
  "info": {
    "title": "Family App API",
    "version": "0.1.0",
    "description": "Family App swagger"
  },
  "host": "family-app-api.herokuapp.com",
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
    "/tasks/{listId}": {
      "get": {
        "operationId": "getAllTasks",
        "summary": "Get all tasks",
        "description": "Get all tasks",
        "produces": "application/json",
        "tags": ["tasks"],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "List Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/GetAllTasks"
            }
          }
        }
      },
      "post": {
        "operationId": "addTask",
        "summary": "add task",
        "description": "add task",
        "produces": "application/json",
        "tags": ["tasks"],
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "description": "List Id"
          },
          {
            "name": "Task JSON",
            "in": "body",
            "description": "Update list body",
            "schema": {
              "$ref": "#/definitions/TaskParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/GetAllTasks"
            }
          }
        }
      }
    },
    "/task/{taskId}": {
      "put": {
        "operationId": "updateTask",
        "summary": "Update Task",
        "description": "Update Task",
        "tags": ["tasks"],
        "produces": "application/json",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task Id"
          },
          {
            "name": "Task JSON",
            "in": "body",
            "description": "Update task body",
            "schema": {
              "$ref": "#/definitions/TaskParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/StatusOk"
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "summary": "Delete Task",
        "tags": ["tasks"],
        "description": "Delete Task",
        "produces": "application/json",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/StatusOk"
            }
          }
        }
      }
    },
    "/task/{taskId}/done": {
      "patch": {
        "operationId": "markTaskAsDone",
        "tags": ["tasks"],
        "summary": "Mark Task as done",
        "description": "Mark Task as done",
        "produces": "application/json",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task Id"
          },
          {
            "name": "Mark Task as Done JSON",
            "in": "body",
            "description": "Change done status: 0 or 1",
            "schema": {
              "$ref": "#/definitions/MarkTaskAsDoneSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/StatusOk"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "TaskParams": {
      "properties": {
        "title": {
          "type": "string"
        },
        "username": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
...
    "TaskEntity": {
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "listId": {
          "type": "string",
          "uniqueItems": true
        },
        "title": {
          "type": "string",
          "uniqueItems": true
        },
        "username": {
          "type": "string",
          "uniqueItems": true
        },
        "done": {
          "type": "number"
        }
      }
    },
...
    "StatusOk": {
      "properties": {
        "status": {}
      }
    },
...
    "MarkTaskAsDoneSchema": {
      "properties": {
        "status": {
          "type": "number"
        }
      }
    },
...
    "GetAllTasks": {
      "type": "array",
      "$ref": "#/definitions/TaskEntity"
    }
  },
...
  "securityDefinitions": {}
}